FEAT Added Reverse, Clear and Date to notes
FEAT Added Reverse, Clear and Date to notes

Reversed the order of the notes that are displayed on the page. Any new
note is added to the beginning of the list.

Clear Button added after the <ul> on the page, which clears out all
notes within localStorage

Added a date value to the notes object. This currently only shows the
Date.now() value when the note is created.

Things still to do:
- We need to convert the date value into a human readable date